Create separate  module for display function and move the function there
